feat: add cleanup cronjob

+++ server/api/src/api/assets/services/cleanup.service.ts
@@ -9,17 +9,20 @@ export class CleanupService {
   constructor(private fileRepository: FileRepository) {}
 
   async cleanupDevice() {
-    // Get all asset_ids that have files with statuses other than processing and failed
-    const assetIdsWithNonProcessingOrFailedFiles = await this.fileRepository.findAssetIdsWithExcludedStatuses([
+    const activeDirectories = await this.fileRepository.findAssetIdsWithStatuses([
       Constants.FILE_STATUS.PROCESSING,
       Constants.FILE_STATUS.QUEUED,
+      Constants.FILE_STATUS.FAILED,
     ]);
 
-    for (const assetId of assetIdsWithNonProcessingOrFailedFiles) {
-      let localPath = Utils.getLocalVideoRootPath(assetId, AppConfigService.appConfig.TEMP_VIDEO_DIRECTORY);
-      console.log('local path ', localPath);
-      if (fs.existsSync(localPath)) {
-        fs.rmSync(localPath, { recursive: true, force: true });
+    let allDirectories = fs.readdirSync(AppConfigService.appConfig.TEMP_VIDEO_DIRECTORY, { withFileTypes: true });
+    for (const dir of allDirectories) {
+      if (dir.isDirectory() && !activeDirectories.includes(dir.name)) {
+        let localPath = Utils.getLocalVideoRootPath(dir.name, AppConfigService.appConfig.TEMP_VIDEO_DIRECTORY);
+        console.log('deleting local path ', localPath);
+        if (fs.existsSync(localPath)) {
+          fs.rmSync(localPath, { recursive: true, force: true });
+        }
       }
     }
   }
